+++ NEWS.md
@@ -28,4 +28,6 @@ Minor issues:
 
 * UI cleanups
 
+* Added a check to verify that ts_ids provided as ordered vectors have the same levels for all provided data frames in ExPanD()
+